+++ storage_utils/tests/test_posix.py
@@ -36,16 +36,25 @@ class TestCopy(unittest.TestCase):
             self.assertTrue(dest.exists())
             self.assertEquals(dest.open().read(), '1')
 
-    def test_ambigious_swift_destination(self):
+    def test_ambigious_swift_resource_destination(self):
         with storage_utils.NamedTemporaryDirectory() as tmp_d:
-            source = tmp_d / 'source'
-            os.mkdir(source)
-            with open(source / '1', 'w') as tmp_file:
+            source = tmp_d / '1'
+            with open(source, 'w') as tmp_file:
                 tmp_file.write('1')
 
-            dest = 'swift://tenant/container/ambiguous'
+            dest = 'swift://tenant/container/ambiguous-resource'
             with self.assertRaisesRegexp(ValueError, 'swift destination'):
-                utils.copy(source / '1', dest)
+                utils.copy(source, dest)
+
+    def test_ambigious_swift_container_destination(self):
+        with storage_utils.NamedTemporaryDirectory() as tmp_d:
+            source = tmp_d / '1'
+            with open(source, 'w') as tmp_file:
+                tmp_file.write('1')
+
+            dest = 'swift://tenant/ambiguous-container'
+            with self.assertRaisesRegexp(ValueError, 'swift destination'):
+                utils.copy(source, dest)
 
     def test_tenant_swift_destination(self):
         with storage_utils.NamedTemporaryDirectory() as tmp_d: